#1834 unit test coverage
