+++ factcast-factus/src/main/java/org/factcast/factus/FactusImpl.java
@@ -25,7 +25,6 @@ import java.lang.reflect.Constructor;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.*;
-import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
@@ -45,7 +44,6 @@ import org.factcast.core.event.EventConverter;
 import org.factcast.core.snap.Snapshot;
 import org.factcast.core.spec.FactSpec;
 import org.factcast.core.subscription.Subscription;
-import org.factcast.core.subscription.SubscriptionClosedException;
 import org.factcast.core.subscription.SubscriptionRequest;
 import org.factcast.core.subscription.observer.FactObserver;
 import org.factcast.factus.batch.DefaultPublishBatch;
@@ -521,42 +519,5 @@ public class FactusImpl implements Factus {
       c.accept(x);
       return x;
     };
-  }
-
-  @RequiredArgsConstructor
-  static class TokenAwareSubscription implements Subscription {
-    final Subscription delegate;
-    final WriterToken token;
-
-    @Override
-    public void close() throws Exception {
-      try {
-        delegate.close();
-      } finally {
-        token.close();
-      }
-    }
-
-    @Override
-    public Subscription awaitCatchup() throws SubscriptionClosedException {
-      return delegate.awaitCatchup();
-    }
-
-    @Override
-    public Subscription awaitCatchup(long waitTimeInMillis)
-        throws SubscriptionClosedException, TimeoutException {
-      return delegate.awaitCatchup(waitTimeInMillis);
-    }
-
-    @Override
-    public Subscription awaitComplete() throws SubscriptionClosedException {
-      return delegate.awaitComplete();
-    }
-
-    @Override
-    public Subscription awaitComplete(long waitTimeInMillis)
-        throws SubscriptionClosedException, TimeoutException {
-      return delegate.awaitComplete(waitTimeInMillis);
-    }
   }
 }
